% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBI.R
\name{CBI}
\alias{CBI}
\title{Cox regression with BART imputation}
\usage{
CBI(
  T,
  C,
  X,
  covariate_types,
  missing_covariate_start = 0,
  time = NULL,
  num_iterations = 1000,
  burn_in = 200,
  beta = NULL,
  h0 = NULL,
  beta_mean = NULL,
  beta_variances = NULL,
  h0_prior_shape = NULL,
  h0_prior_rate = NULL,
  m = 200,
  numcut = 100,
  alpha = 0.95,
  mybeta = 2,
  nu = 3,
  lambda = 1,
  proposal_stddev_beta = 1,
  proposal_stddev_h0 = 1,
  proposal_stddev_X = 1
)
}
\arguments{
\item{C}{Censoring indicators}

\item{X}{Covariates}

\item{covariate_types}{Covariate types (1: continuous, 2: binary)}

\item{missing_covariate_start}{Index of the first covariate with missing values}

\item{time}{Piecewise time points for the baseline hazard}

\item{num_iterations}{Number of iterations for MCMC}

\item{burn_in}{Number of burn-in iterations}

\item{beta}{Initial beta values}

\item{h0}{Initial baseline hazard}

\item{beta_mean}{Mean of the beta prior}

\item{beta_variances}{Variances of the beta prior}

\item{h0_prior_shape}{Shape of the baseline hazard prior}

\item{h0_prior_rate}{Rate of the baseline hazard prior}

\item{m}{Number of trees in the BART model}

\item{numcut}{Number of cutpoints for splitting in the BART model}

\item{alpha}{Parameter for the depth of the trees in BART (p = alpha * (1 + d)^mybeta)}

\item{mybeta}{Parameter for the number of terminal nodes in BART  (p = alpha * (1 + d)^mybeta)}

\item{nu}{Degrees of freedom for sigma prior}

\item{lambda}{Scale for sigma prior}

\item{proposal_stddev_beta}{Standard deviation for the proposal distribution of beta}

\item{proposal_stddev_h0}{Standard deviation for the proposal distribution of h0}

\item{proposal_stddev_X}{Standard deviation for the proposal distribution of X}
}
\value{
A list with the following components: A matrix of beta samples, last imputed X
}
\description{
Cox regression with BART imputation
}
\examples{
X = test_data(1, n = 500)
T = X$observe
C = X$status
X = X[, -c(1, 2)]
covariate_types = c(1, 2, 1)
missing_covariate_start = 2
X = as.matrix(X)
result = CBI(T, C, X, covariate_types, missing_covariate_start, num_iterations = 10000,burn_in = 5000)
}
