% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{test_data}
\alias{test_data}
\title{Generate a Weibull survival dataset with missing covariates by MAR / MCAR}
\usage{
test_data(
  seed,
  n = 500,
  lambda = 0.2,
  K = 1.5,
  q = 0.2,
  mis = 0.2,
  method = "MCAR"
)
}
\arguments{
\item{seed}{Set seed}

\item{n}{Number of observations}

\item{lambda}{Scale parameter of the Weibull distribution}

\item{K}{Shape parameter of the Weibull distribution}

\item{q}{Censoring proportion}

\item{mis}{Missing proportion}

\item{method}{Missing data mechanism: "MCAR" or "MAR"}
}
\value{
A data frame with the following columns: observe, status, x1, x2, x3
}
\description{
Generate a Weibull survival dataset with missing covariates by MAR / MCAR
}
\examples{
data = test_data(1, n = 500)
data = test_data(1, n = 500, lambda = 0.1, K = 1.5, q = 0.2)
}
